Proto major/minors shouldn't live in the extension header but the proto header
Previously the Xserver was pulling in the extension header rather than the proto header
this led to issues when compiling the xserver. Instead we shift the proto numbers to
the proto header and tell the X server to use the proto header. Hence it's possible
to compile the server without the lib again.

Signed-off-by: Benjamin Close <Benjamin.Close@clearchain.com>
Acked-by: Daniel Stone <daniel@fooishbar.org>
Acked-by: Peter Hutterer <peter.hutterer@who-t.net>