libXinerama: Set number of screens to 0 on error
The documentation of XineramaQueryScreens states that number is always
set to the amount of screens in the returned array, but if the
communication with the X server fails, NULL is returned without
modifying number.

At least dwm relies on the fact that number is set to 0 on error, i.e.
when NULL is returned. As a NULL pointer contains 0 elements and the
documentation states that number contains the amount of elements of the
returned array, I think this should be fixed inside libXinerama.

Reviewed-by: Adam Jackson <ajax@redhat.com>